+++ lib/async/bus/server.rb
@@ -20,13 +20,23 @@ module Async
 				@options = options
 			end
 			
+			# Called when a connection is established.
+			# Override this method to perform setup when a connection is established.
+			#
+			# @parameter connection [Protocol::Connection] The established connection.
+			protected def connected!(connection)
+				# Do nothing by default.
+			end
+			
 			# Accept incoming connections.
 			# @yields {|connection| ...} Block called with each new connection.
-			def accept
+			def accept(&block)
 				@endpoint.accept do |peer|
 					connection = Protocol::Connection.server(peer, **@options)
 					
-					yield connection
+					connected!(connection, &block)
+					
+					yield connection if block_given?
 					
 					connection.run
 				ensure