+++ releases.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
   - Add support for multi-hop proxying.
+  - Fix proxying of throw/catch value.
+  - `Client#run` now takes a block.
+  - `Server#run` delegates to `Server#connected!`.
 
 ## v0.2.0